Struct sc_client_db::Backend

pub struct Backend<Block: BlockT> { /* fields omitted */ }

Disk backend.

Disk backend keeps data in a key-value store. In archive mode, trie nodes are kept from all blocks. Otherwise, trie nodes are kept only from some recent blocks.

Implementations

impl<Block: BlockT> Backend<Block>

pub fn new(
config: DatabaseSettings,
canonicalization_delay: u64
) -> ClientResult<Self>

Create a new instance of database backend.

The pruning window is how old a block must be before the state is pruned.