Struct sc_client_db::RefTrackingState

pub struct RefTrackingState<Block: BlockT> { /* private fields */ }

A reference tracking state.

It makes sure that the hash we are using stays pinned in storage until this structure is dropped.