Trait sc_client_db::Database

pub trait Database<H>: Send + Sync where
    H: Clone + AsRef<[u8]>, {
    pub fn commit(
        &self, 
        transaction: Transaction<H>
    ) -> Result<(), DatabaseError>;

    pub fn get(&self, col: u32, key: &[u8]) -> Option<Vec<u8, Global>>;

    pub fn contains(&self, col: u32, key: &[u8]) -> bool { ... }

    pub fn value_size(&self, col: u32, key: &[u8]) -> Option<usize> { ... }

    pub fn with_get(&self, col: u32, key: &[u8], f: &mut dyn FnMut(&[u8])) { ... }
}

Required methods

pub fn commit(&self, transaction: Transaction<H>) -> Result<(), DatabaseError>

Commit the transaction to the database atomically. Any further calls to get or lookup will reflect the new state.

pub fn get(&self, col: u32, key: &[u8]) -> Option<Vec<u8, Global>>

Retrieve the value previously stored against key or None if key is not currently in the database.

Loading content...
Provided methods

pub fn contains(&self, col: u32, key: &[u8]) -> bool

Check if the value exists in the database without retrieving it.

pub fn value_size(&self, col: u32, key: &[u8]) -> Option<usize>

Check value size in the database possibly without retrieving it.

pub fn with_get(&self, col: u32, key: &[u8], f: &mut dyn FnMut(&[u8]))

Call f with the value previously stored against key.

This may be faster than get since it doesn’t allocate. Use with_get helper function if you need f to return a value from f

Loading content...