Crate sc_client_db

Client backend that is backed by a database.

Canonicality vs. Finality

Finality indicates that a block will not be reverted, according to the consensus algorithm, while canonicality indicates that the block may be reverted, but we will be unable to do so, having discarded heavy state that will allow a chain reorganization.

Finality implies canonicality but not vice-versa.

Re-exports

pub use bench::BenchmarkingState;

Modules

bench

State backend that’s useful for benchmarking

offchain

RocksDB-based offchain workers local storage.

Structs

Backend

Disk backend.

BlockImportOperation

Database transaction

BlockchainDb

Block database

DatabaseSettings

Database settings.

RefTrackingState

A reference tracking state.

Enums

DatabaseSource

Where to find the database..

KeepBlocks

Block pruning settings.

PruningMode

Pruning mode.

Traits

Database

Type Definitions

DbHash

Hash type that this backend uses for the database.

DbState

DB-backed patricia trie state, transaction type is an overlay of changes to commit.