Struct sc_client_db::Backend

pub struct Backend<Block: BlockT> { /* private fields */ }

Disk backend.

Disk backend keeps data in a key-value store. In archive mode, trie nodes are kept from all blocks. Otherwise, trie nodes are kept only from some recent blocks.

Implementations

impl<Block: BlockT> Backend<Block>

pub fn new( db_config: DatabaseSettings, canonicalization_delay: u64, ) -> ClientResult<Self>

Create a new instance of database backend.

The pruning window is how old a block must be before the state is pruned.

pub fn reset_trie_cache(&self)

Reset the shared trie cache.

pub fn new_test(blocks_pruning: u32, canonicalization_delay: u64) -> Self

Create new memory-backed client backend for tests.

pub fn new_test_with_tx_storage( blocks_pruning: BlocksPruning, canonicalization_delay: u64, ) -> Self

Create new memory-backed client backend for tests.

pub fn expose_db(&self) -> (Arc<dyn Database<DbHash>>, ColumnId)

Expose the Database that is used by this backend. The second argument is the Column that stores the State.

Should only be needed for benchmarking.

pub fn expose_storage(&self) -> Arc<dyn Storage<HashingFor<Block>>>

Expose the Storage that is used by this backend.

Should only be needed for benchmarking.